Fix nasty circular dependency
